new stable update with working fix

+++ src/Recode.Service/Implementations/EntityService/UserService.cs
@@ -24,18 +24,30 @@ namespace Recode.Service.EntityService
         private readonly IRepositoryQuery<User, long> _userQueryRepo;
         private readonly IRepositoryCommand<UserRole, long> _userRoleCommandRepo;
         private readonly IRepositoryQuery<Role, long> _roleQueryRepo;
+        private readonly IRepositoryQuery<Company, long> _companyQueryRepo;
         private readonly IEmailService _emailService;
         private readonly IMapper _mapper;
         private readonly IRepositoryQuery<UserRole, long> _userRoleQueryRepo;
         private readonly SSoSetting _ssoSettings;
 
         private readonly string CurrentUserId;
+        private long _currentCompanyId;
+        public long CurrentCompanyId
+        {
+            get
+            {
+                _currentCompanyId = _currentCompanyId == 0 ? _httpContext.GetCurrentCompanyId() : _currentCompanyId;
+                return _currentCompanyId;
+            }
+        }
+
         public UserService(ISSOService ssoService,
             IRepositoryCommand<User, long> userCommandRepo,
             IRepositoryQuery<User, long> userQueryRepo,
             IRepositoryCommand<UserRole, long> userRoleCommandRepo,
             IRepositoryQuery<UserRole, long> userRoleQueryRepo,
             IRepositoryQuery<Role, long> roleQueryRepo,
+            IRepositoryQuery<Company, long> companyQueryRepo,
             IEmailService emailService,
             IMapper mapper,
              IOptions<SSoSetting> sSoSetting, IHttpContextExtensionService httpContext)
@@ -46,32 +58,94 @@ namespace Recode.Service.EntityService
             _userQueryRepo = userQueryRepo;
             _userRoleCommandRepo = userRoleCommandRepo;
             _roleQueryRepo = roleQueryRepo;
+            _companyQueryRepo = companyQueryRepo;
             _emailService = emailService;
             _mapper = mapper;
             _userRoleQueryRepo = userRoleQueryRepo;
             _httpContext = httpContext;
 
-            CurrentUserId = string.IsNullOrEmpty(_httpContext.GetCurrentUserId()) ? "testuser" : _httpContext.GetCurrentUserId();
+            CurrentUserId = _httpContext.GetCurrentUserId();
         }
 
         public async Task<ExecutionResponse<UserModel>> AddUserRole(UserRoleModel model)
         {
-            await _userRoleCommandRepo.InsertAsync(new UserRole { CreateById = CurrentUserId, RoleId = model.RoleId, UserId = model.UserId });
-            await _userRoleCommandRepo.SaveChangesAsync();
-            return await GetUser(model.UserId);
+            var role = _roleQueryRepo.GetAll().FirstOrDefault(f => f.Id == model.RoleId);
+            if (role == null)
+            {
+                return new ExecutionResponse<UserModel>
+                {
+                    ResponseCode = ResponseCode.NotFound,
+                    Message = "Role does not exist"
+                };
+            }
+            
+            if (CurrentCompanyId == 0 && role.RoleName != "CompanyAdmin")
+                throw new Exception($"You cannot add role '{role.RoleName}' to a user");
+
+            var user = _userQueryRepo.GetAll().FirstOrDefault(f => f.Id == model.UserId);
+            if (user == null || (CurrentCompanyId != 0 && user.CompanyId != CurrentCompanyId))
+            {
+                return new ExecutionResponse<UserModel>
+                {
+                    ResponseCode = ResponseCode.NotFound,
+                    Message = "User does not exist"
+                };
+            }
+
+            if (!user.UserRoles.Any(u => u.RoleId == role.Id))
+            {
+                var result = await _ssoService.AddRemoveClaims(new UserClaimModel { Claims = new List<SSOClaim> { new SSOClaim("role", role.RoleName) }, UserId = user.SSOUserId }, ClaimAction.Add);
+
+                if (result.ResponseCode == ResponseCode.Ok)
+                {
+                    await _userRoleCommandRepo.InsertAsync(new UserRole { RoleId = role.Id, UserId = user.Id });
+                    await _userRoleCommandRepo.SaveChangesAsync();
+
+                    return await GetUser(model.UserId);
+                }
+                else
+                {
+                    return new ExecutionResponse<UserModel>
+                    {
+                        ResponseCode = ResponseCode.ServerException,
+                        Message = result.Message
+                    };
+                }
+            }
+            else
+            {
+                return new ExecutionResponse<UserModel>
+                {
+                    ResponseCode = ResponseCode.ServerException,
+                    Message = "User is already attached to specified role"
+                };
+            }
         }
 
         public async Task<ExecutionResponse<UserModel>> CreateUser(CreateUserModel model)
         {
-            var role = _roleQueryRepo.GetAll().FirstOrDefault(x => x.Id == model.RoleId);
+            if (_userQueryRepo.GetAll().Any(u => u.Email.Trim().ToLower() == model.Email.Trim().ToLower()))
+                return new ExecutionResponse<UserModel>
+                {
+                    ResponseCode = ResponseCode.ServerException,
+                    Message = "User with specified email already exists"
+                };
 
+            var role = _roleQueryRepo.GetAll().FirstOrDefault(x => x.Id == model.RoleId );
             if (role == null)
                 throw new Exception("Invalid Role");
 
+            if (CurrentCompanyId == 0 && role.RoleName != "CompanyAdmin")
+                throw new Exception($"You cannot create a user with role '{role.RoleName}'");
+
+            var company = _companyQueryRepo.GetAll().FirstOrDefault(x => x.Id == model.CompanyId);
+            if (company == null)
+                throw new Exception("Company does not exist");
+
             var registerReponse = await _ssoService.Register(new SSOUserDto
             {
-                Claims = new List<Claim> { new Claim("role", GetSSORole(role.RoleName)),
-                                            new Claim("role", role.RoleName) },
+                Claims = new List<SSOClaim> { new SSOClaim("role", GetSSORole(role.RoleName)),
+                                            new SSOClaim("role", role.RoleName) },
                 ConfirmEmail = false,
                 Email = model.Email,
                 FirstName = model.FirstName,
@@ -89,6 +163,8 @@ namespace Recode.Service.EntityService
                 };
 
             //user registered on sso 
+            //trigger reset password
+
             //save user info
             var user = new User
             {
@@ -97,7 +173,7 @@ namespace Recode.Service.EntityService
                 LastName = model.LastName,
                 PhoneNumber = string.IsNullOrEmpty(model.PhoneNumber) ? "08111111111" : model.PhoneNumber,
                 UserName = model.UserName,
-                CompanyId = model.CompanyId,
+                CompanyId = CurrentCompanyId == 0 ? model.CompanyId : CurrentCompanyId,
                 EmailConfirmed = false,
                 CreateById = CurrentUserId,
                 SSOUserId = registerReponse.ResponseData.UserId
@@ -106,7 +182,7 @@ namespace Recode.Service.EntityService
             await _userCommandRepo.SaveChangesAsync();
 
             //assign role to user
-            await _userRoleCommandRepo.InsertAsync(new UserRole {RoleId = role.Id, UserId = user.Id, CreateById = CurrentUserId });
+            await _userRoleCommandRepo.InsertAsync(new UserRole { RoleId = role.Id, UserId = user.Id, CreateById = CurrentUserId });
             await _userRoleCommandRepo.SaveChangesAsync();
 
             //send email confirmation mail
@@ -122,7 +198,7 @@ namespace Recode.Service.EntityService
 
         private string GetSSORole(string roleName)
         {
-           switch (roleName)
+            switch (roleName)
             {
                 case "VGG_Admin":
                     return "vgg_admin";
@@ -155,7 +231,7 @@ namespace Recode.Service.EntityService
         {
             var user = _userQueryRepo.GetAll().FirstOrDefault(x => x.Id == Id);
 
-            if (user == null)
+            if (user == null || (CurrentCompanyId != 0 && user.CompanyId != CurrentCompanyId))
                 return new ExecutionResponse<UserModel>
                 {
                     ResponseCode = ResponseCode.NotFound,
@@ -173,7 +249,7 @@ namespace Recode.Service.EntityService
         {
             var user = _userQueryRepo.GetAll().FirstOrDefault(x => x.Email.Trim().ToLower() == email.Trim().ToLower());
 
-            if (user == null)
+            if (user == null || (CurrentCompanyId != 0 && user.CompanyId != CurrentCompanyId))
                 return new ExecutionResponse<UserModel>
                 {
                     ResponseCode = ResponseCode.NotFound,
@@ -189,19 +265,23 @@ namespace Recode.Service.EntityService
 
         public async Task<ExecutionResponse<UserModel[]>> GetUsersByRoleId(long roleId)
         {
-            var userRoles = _userRoleQueryRepo.GetAll().Include(x=>x.User).Where(x => x.RoleId == roleId)
-                            .Select(x => new[] { x.User }).ToList();
+            var userRoles = _userRoleQueryRepo.GetAll().Include(x => x.User).Where(x => x.RoleId == roleId)
+                            .Select(x => _mapper.Map<UserModel>(x.User)).ToArray();
+
+            userRoles = CurrentCompanyId == 0 ? userRoles : userRoles.Where(u => u.CompanyId == CurrentCompanyId).ToArray();
 
             return new ExecutionResponse<UserModel[]>
             {
                 ResponseCode = ResponseCode.Ok,
-                ResponseData = _mapper.Map<UserModel[]>(userRoles)
+                ResponseData = userRoles
             };
-        }        
+        }
 
         public async Task<ExecutionResponse<UserModel[]>> GetUsers(long[] userIds)
         {
-            var users = _userQueryRepo.GetAll().Where(x => userIds.Contains(x.Id));
+            var users = _userQueryRepo.GetAll().Where(x => userIds.Contains(x.Id)).ToArray();
+
+            users = CurrentCompanyId == 0 ? users : users.Where(u => u.CompanyId == CurrentCompanyId).ToArray();
 
             return new ExecutionResponse<UserModel[]>
             {
@@ -214,7 +294,7 @@ namespace Recode.Service.EntityService
         {
             var users = _userQueryRepo.GetAll();//.Where(x => x.CompanyId == _httpContext.GetCurrentCompanyId());
 
-            users = _httpContext.GetCurrentCompanyId() == 0 ? users : users.Where(x => x.CompanyId == _httpContext.GetCurrentCompanyId());
+            users = CurrentCompanyId == 0 ? users : users.Where(x => x.CompanyId == CurrentCompanyId);
             users = string.IsNullOrEmpty(email) ? users : users.Where(x => x.Email.Contains(email));
             users = string.IsNullOrEmpty(firstName) ? users : users.Where(x => x.FirstName.Contains(firstName));
             users = string.IsNullOrEmpty(lastName) ? users : users.Where(x => x.LastName.Contains(lastName));
@@ -240,17 +320,60 @@ namespace Recode.Service.EntityService
         {
             var userRole = _userRoleQueryRepo.GetAll().FirstOrDefault(x => x.UserId == model.UserId && x.RoleId == model.RoleId);
             if (userRole != null)
-                await _userRoleCommandRepo.DeleteAsync(userRole.Id);
-            await _userRoleCommandRepo.SaveChangesAsync();
+            {
+                var role = _roleQueryRepo.GetAll().FirstOrDefault(f => f.Id == model.RoleId);
+                if (role == null)
+                {
+                    return new ExecutionResponse<UserModel>
+                    {
+                        ResponseCode = ResponseCode.NotFound,
+                        Message = "Role does not exist"
+                    };
+                }
 
-            return await GetUser(model.UserId);
+                var user = _userQueryRepo.GetAll().FirstOrDefault(f => f.Id == model.UserId);
+                if (user == null || (CurrentCompanyId != 0 && user.CompanyId != CurrentCompanyId))
+                {
+                    return new ExecutionResponse<UserModel>
+                    {
+                        ResponseCode = ResponseCode.NotFound,
+                        Message = "User does not exist"
+                    };
+                }
+
+                var result = await _ssoService.AddRemoveClaims(new UserClaimModel { Claims = new List<SSOClaim> { new SSOClaim("role", role.RoleName) }, UserId = user.SSOUserId }, ClaimAction.Remove);
+
+                if (result.ResponseCode == ResponseCode.Ok)
+                {
+                    await _userRoleCommandRepo.DeleteAsync(userRole.Id);
+                    await _userRoleCommandRepo.SaveChangesAsync();
+
+                    return await GetUser(model.UserId);
+                }
+                else
+                {
+                    return new ExecutionResponse<UserModel>
+                    {
+                        ResponseCode = ResponseCode.ServerException,
+                        Message = result.Message
+                    };
+                }
+            }
+            else
+            {
+                return new ExecutionResponse<UserModel>
+                {
+                    ResponseCode = ResponseCode.ServerException,
+                    Message = "User is not attached to specified role"
+                };
+            }
         }
 
         public async Task<ExecutionResponse<UserModel>> ToggleActivateUser(long Id)
         {
-             var user = _userQueryRepo.GetAll().FirstOrDefault(x => x.Id == Id);
+            var user = _userQueryRepo.GetAllIncludeInactive().FirstOrDefault(x => x.Id == Id);
 
-            if (user == null)
+            if (user == null || (CurrentCompanyId != 0 && user.CompanyId != CurrentCompanyId))
                 return new ExecutionResponse<UserModel>
                 {
                     ResponseCode = ResponseCode.NotFound,
@@ -273,7 +396,7 @@ namespace Recode.Service.EntityService
         {
             var user = _userQueryRepo.GetAll().FirstOrDefault(x => x.Id == model.Id);
 
-            if (user == null)
+            if (user == null || (CurrentCompanyId != 0 && user.CompanyId != CurrentCompanyId))
                 return new ExecutionResponse<UserModel>
                 {
                     ResponseCode = ResponseCode.NotFound,
@@ -324,7 +447,50 @@ namespace Recode.Service.EntityService
             if (ssoresponse.ResponseCode != ResponseCode.Ok)
                 return false;
 
-           return _emailService.EmailConfirmation(ssoresponse.ResponseData.EmailConfirmationToken, user.Email, $"{user.FirstName} {user.LastName}", user.SSOUserId);
+            return _emailService.EmailConfirmation(ssoresponse.ResponseData.EmailConfirmationToken, user.Email, $"{user.FirstName} {user.LastName}", user.SSOUserId);
+        }
+
+        public async Task<ExecutionResponse<UserModel>> OnboardAdmin(OnboardUserModel model)
+        {
+            var ssoUser = await _ssoService.GetUser(model.Id);
+
+            if (ssoUser.ResponseCode != ResponseCode.Ok)
+            {
+                return new ExecutionResponse<UserModel>
+                {
+
+                };
+            }
+
+            var user = new User
+            {
+                SSOUserId = model.Id,
+                FirstName = ssoUser.ResponseData.Response.FirstName,
+                LastName = ssoUser.ResponseData.Response.LastName,
+                UserName = ssoUser.ResponseData.Response.UserName,
+                Email = ssoUser.ResponseData.Response.Email,
+                CreateById = "VGGSuperAdmin",
+                PhoneNumber = ssoUser.ResponseData.Response.PhoneNumber,
+                EmailConfirmed = ssoUser.ResponseData.EmailConfirmed,
+                CompanyId = 0
+            };
+
+            await _userCommandRepo.InsertAsync(user);
+            await _userCommandRepo.SaveChangesAsync();
+
+            var role = _roleQueryRepo.GetAll().FirstOrDefault(r => r.RoleName == "VGG_Admin");
+
+            if (role != null)
+            {
+                await _userRoleCommandRepo.InsertAsync(new UserRole { RoleId = role.Id, UserId = user.Id, CreateById = "VGGSuperAdmin" });
+                await _userRoleCommandRepo.SaveChangesAsync();
+            }
+
+            return new ExecutionResponse<UserModel>
+            {
+                ResponseCode = ResponseCode.Ok,
+                ResponseData = _mapper.Map<UserModel>(user)
+            };
         }
     }
 }
